requirements.txt
google-genai
python-pptx
python-dotenv
lxml
flask
flask-cors
groq
gunicorn